Draw Page with meta tag for search engine, and Facebook Graph better recognition

+++ draw/delete.php
@@ -2,6 +2,6 @@
 include '../connect.php';
 include '../session_check.php';
 $id=$_POST['id'];
-$q = mysqli_query($link,"delete from draw where id = $id")or die(mysqli_error($link));
+$q = mysqli_query($link,"delete from draw where id = $id and userid = $userid")or die(mysqli_error($link));
 echo 1;
 ?>
